requirements.txt
pytz
kafka-python
SQLAlchemy
Flask
flask-socketio
flask-cors
